lab2/lab5_project1/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
bidict==0.23.1
certifi==2024.6.2
charset-normalizer==3.3.2
dnspython==2.6.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
loguru==0.7.2
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
python-engineio==4.9.1
python-socketio==5.11.3
requests==2.32.3
simple-websocket==1.0.0
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
